// src/core/utils/enums.ts
export enum MessageType {
    Text = "text",
    Voice = "voice",
    Image = "image",
    Video = "video",
    File = "file",
    AllDeleted = "allDeleted",
    Location = "location",
    Custom = "custom",
    Call = "call",
    //this should not notify user about it
    Info = "info",
}
// ...
export enum SocketEventsType {
    //clint listener
    v1OnNewMessage = "v1OnNewMessage",
    v1OnBanUserChat = "v1OnBanUserChat",
    v1OnDeleteMessageFromAll = "v1OnDeleteMessageFromAll",
    v1OnKickGroupMember = "v1OnKickGroupMember",
    v1OnMyOnline = "v1OnMyOnline",
    v1OnRoomStatusChange = "v1OnRoomStatusChange",
    v1OnDeliverChatRoom = "v1OnDeliverChatRoom",
    v1OnException = "v1OnException",
    v1OnEnterChatRoom = "v1OnEnterChatRoom",
    //call events
    v1OnCallAccepted = "v1OnCallAccepted",
    v1OnCallEnded = "v1OnCallEnded",
    v1OnCallCanceled = "v1OnCallCanceled",
    v1OnCallRejected = "v1OnCallRejected",
    v1OnNewCall = "v1OnNewCall",
    v1OnCallTimeout = "v1OnCallTimeout",
    v1OnIceCandidate = "v1OnIceCandidate",


    //server emitter
    v1EnterChatRoom = "v1EnterChatRoom",
    v1KickGroupMember = "v1KickGroupMember",
    v1DeliverChatRoom = "v1DeliverChatRoom",
    v1MyOnline = "v1MyOnline",
    v1RoomStatusChange = "v1RoomStatusChange",
    v1IceCandidate = "v1IceCandidate",
}
// ...
export enum CallStatus {
    Ring = "ring",
    Canceled = "canceled",
    Timeout = "timeout",
    Rejected = "rejected",
    Finished = "finished",
    SessionEnd = "sessionEnd",
    InCall = "inCall",
}